Make frame detection more robust
After we made the Widget Preview sidebar icon appear immediately, we would click and find an iframe, but that iframe would then be replaced (apparently) when the view was resolved, resulting in "locator.waitFor: Frame was detached".

Now we'll check for an embedded Flutter app instead (since our nested views are always Flutter apps) so we never proceed with that fake frame.